+++ app/definitions/pages/common/file-upload/getting-digital-receipts-or-invoices.js
@@ -1,4 +1,5 @@
 const fieldValidators = require('../../../field-validators/common/optional-validator');
+const { claimTypesShortName } = require('../../../../config/claim-types');
 
 module.exports = () => ({
   view: 'pages/common/file-upload/getting-digital-receipts-or-invoices.njk',
@@ -6,6 +7,7 @@ module.exports = () => ({
   hooks: {
     prerender: (req, res, next) => {
       res.locals.journeyType = req.casa.journeyContext.getDataForPage('__journey_type__').journeyType;
+      res.locals.awardType = claimTypesShortName[res.locals.journeyType];
       next();
     },
   },